Moved entry_list under managed_lists component
