+++ flexget/components/managed_lists/lists/entry_list/api.py
