+++ flexget/components/managed_lists/lists/entry_list/entry_list.py
