+++ flexget/tests/api_tests/test_entry_list_api.py
@@ -4,10 +4,10 @@ from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 import copy
 
 from flexget.api.app import base_message
-from flexget.components.entry_list.api import ObjectsContainer as OC
+from flexget.components.managed_lists.lists.entry_list.api import ObjectsContainer as OC
 from flexget.entry import Entry
 from flexget.manager import Session
-from flexget.components.entry_list.db import EntryListList, EntryListEntry
+from flexget.components.managed_lists.lists.entry_list.db import EntryListList, EntryListEntry
 from flexget.utils import json
